backend/cart_service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
sqlalchemy==2.0.21
alembic==1.12.0
asyncpg==0.28.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
httpx==0.25.0
python-dotenv==1.0.0
pyjwt==2.8.0 